data-analyzer/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.7
pandas==2.1.3
python-dotenv==1.0.0
